segmentation/pyproject.toml
[tool.poetry]
name = "awesome-semantic-segmentation-pytorch"
version = "0.1.0"
description = ""
authors = ["Mo"]

[tool.poetry.dependencies]
python = "^3.7"
ninja = "^1.9.0"
tqdm = "^4.41.1"
numpy = "^1.18.1"
requests = "^2.22.0"
scipy = "^1.4.1"
pillow = "6.1"
torch = "^1.4.0"
torchvision = "^0.5.0"
tensorboard = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
